adding controllers&routes

+++ src/services/animal.service.js
@@ -11,24 +11,24 @@ class AnimalService{
         let animal = await Animal.create(value)
         return animal
     }
-    static async getAnimalById(id){
-        let animal = await Animal.findOne({where:{id:id}})
+    static async getAnimalById(id,vetId){
+        let animal = await Animal.findOne({where:{id:id,createdBy:vetId}})
         return animal
 
     }
-    static async getAllanimals(){
-        let animal = await Animal.findAll()
+    static async getAllanimals(vetId){
+        let animal = await Animal.findAll({where:{createdBy:vetId}})
         return animal
 
     }
     static async updateById(id,value){
-        let animal = await Animal.update(value,{where:{id:id}})
+        let animal = await Animal.update(value,{where:{id:id,createdBy:vetId}})
         return animal
 
     }
 
-    static async deleteById(id){
-        let animal = await Animal.destroy({where:{id:id}})
+    static async deleteById(id,vetId){
+        let animal = await Animal.destroy({where:{id:id,createdBy:vetId}})
         return animal
     }
